Update README.md
# My Building AI Project

## Summary
This project was created as part of the **Building AI** course by the **University of Helsinki**.  
It demonstrates a simple idea of how artificial intelligence can be used to help people make smarter decisions in daily life.  
The goal is to use data and algorithms to create a model that predicts outcomes and supports learning.  

## Background
I took this course to understand how AI works in practice and to build my first project using what I learned.  
This repository is part of my final submission for the course.

## Acknowledgement
Thanks to **Elements of AI** and **Reaktor** for making this course freely available and inspiring learners around the world.